// src/app/pages/clientes/components/clientes-table/clientes.interface.ts
export interface ClientesInterface {
    claveauth?: string;
    nicknameauth?: string;
    usuarioauth?: string;
    idcliente?: number;
    descripcion: string;
    direccion: string;
    medidasterreno: string;
    medidasconstruccion: string;
    fechainicio: string;
    fechafin: string;
    idtipocliente: number;
    presupuesto: number;
    idrazonsocialcliente: number;
    idrazonsocialcontratista: number;
    idrazonsocialconstructor: number;
    idrazonsocialasociado: number;
    posiciongps: string;
    idestatuscliente: number;
    observaciones: string;
    desctipocliente?: string;
    direccioncliente?: string;
    razonsocialcontratista?: string;
    razonsocialconstructor?: string;
    razonsocialasociado?: string;
    claveestatuscliente?: string;
    tipocliente?: string;
}